+++ server/lib/nl/fulfillment/comparison.py
@@ -15,6 +15,9 @@
 import logging
 from typing import List
 
+from server.lib.nl.common.utterance import ChartOriginType
+from server.lib.nl.common.utterance import ChartType
+from server.lib.nl.common.utterance import Utterance
 from server.lib.nl.detection.types import Place
 from server.lib.nl.fulfillment.base import add_chart_to_utterance
 from server.lib.nl.fulfillment.base import populate_charts_for_places
@@ -22,9 +25,6 @@ from server.lib.nl.fulfillment.context import \
     places_for_comparison_from_context
 from server.lib.nl.fulfillment.types import ChartVars
 from server.lib.nl.fulfillment.types import PopulateState
-from server.lib.nl.utterance import ChartOriginType
-from server.lib.nl.utterance import ChartType
-from server.lib.nl.utterance import Utterance
 
 
 def populate(uttr: Utterance) -> bool:
