+++ server/lib/nl/fulfillment/containedin.py
@@ -15,7 +15,10 @@
 import logging
 from typing import List
 
-from server.lib.nl import utils
+from server.lib.nl.common import utils
+from server.lib.nl.common.utterance import ChartOriginType
+from server.lib.nl.common.utterance import ChartType
+from server.lib.nl.common.utterance import Utterance
 from server.lib.nl.detection.types import ClassificationType
 from server.lib.nl.detection.types import ContainedInClassificationAttributes
 from server.lib.nl.detection.types import Place
@@ -25,9 +28,6 @@ from server.lib.nl.fulfillment.context import \
     classifications_of_type_from_context
 from server.lib.nl.fulfillment.types import ChartVars
 from server.lib.nl.fulfillment.types import PopulateState
-from server.lib.nl.utterance import ChartOriginType
-from server.lib.nl.utterance import ChartType
-from server.lib.nl.utterance import Utterance
 
 
 def populate(uttr: Utterance) -> bool:
