+++ server/lib/nl/fulfillment/context.py
@@ -14,13 +14,13 @@
 
 from typing import Dict, List
 
-from server.lib.nl import constants
+from server.lib.nl.common import constants
+from server.lib.nl.common.utterance import CTX_LOOKBACK_LIMIT
+from server.lib.nl.common.utterance import QueryType
+from server.lib.nl.common.utterance import Utterance
 from server.lib.nl.detection.types import ClassificationAttributes
 from server.lib.nl.detection.types import ClassificationType
 from server.lib.nl.detection.types import Place
-from server.lib.nl.utterance import CTX_LOOKBACK_LIMIT
-from server.lib.nl.utterance import QueryType
-from server.lib.nl.utterance import Utterance
 
 #
 # General utilities for retrieving stuff from past context.
