+++ server/lib/nl/fulfillment/correlation.py
@@ -15,7 +15,11 @@
 import logging
 from typing import List
 
-from server.lib.nl import utils
+from server.lib.nl.common import utils
+from server.lib.nl.common.topic import open_top_topics_ordered
+from server.lib.nl.common.utterance import ChartOriginType
+from server.lib.nl.common.utterance import ChartType
+from server.lib.nl.common.utterance import Utterance
 from server.lib.nl.detection.types import ClassificationType
 from server.lib.nl.detection.types import ContainedInClassificationAttributes
 from server.lib.nl.detection.types import ContainedInPlaceType
@@ -29,10 +33,6 @@ from server.lib.nl.fulfillment.context import places_from_context
 from server.lib.nl.fulfillment.context import svs_from_context
 from server.lib.nl.fulfillment.types import ChartVars
 from server.lib.nl.fulfillment.types import PopulateState
-from server.lib.nl.topic import open_top_topics_ordered
-from server.lib.nl.utterance import ChartOriginType
-from server.lib.nl.utterance import ChartType
-from server.lib.nl.utterance import Utterance
 import shared.lib.detected_variables as vars
 
 _MAX_CONTEXT_SVS = 3
