+++ server/lib/nl/fulfillment/ranking_across_vars.py
@@ -15,16 +15,16 @@
 import logging
 from typing import List
 
-from server.lib.nl import rank_utils
-from server.lib.nl import utils
+from server.lib.nl.common import rank_utils
+from server.lib.nl.common import utils
+from server.lib.nl.common.utterance import ChartOriginType
+from server.lib.nl.common.utterance import ChartType
+from server.lib.nl.common.utterance import Utterance
 from server.lib.nl.detection.types import Place
 from server.lib.nl.fulfillment.base import add_chart_to_utterance
 from server.lib.nl.fulfillment.base import populate_charts
 from server.lib.nl.fulfillment.types import ChartVars
 from server.lib.nl.fulfillment.types import PopulateState
-from server.lib.nl.utterance import ChartOriginType
-from server.lib.nl.utterance import ChartType
-from server.lib.nl.utterance import Utterance
 
 _MAX_VARS_IN_A_CHART = 20
